+++ src/onegov/org/locale/de_CH/LC_MESSAGES/onegov.org.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE 1.0\n"
-"POT-Creation-Date: 2024-04-02 09:42+0200\n"
+"POT-Creation-Date: 2024-04-30 08:47+0200\n"
 "PO-Revision-Date: 2022-03-15 10:21+0100\n"
 "Last-Translator: Marc Sommerhalder <marc.sommerhalder@seantis.ch>\n"
 "Language-Team: German\n"
@@ -1013,6 +1013,17 @@ msgstr "Ein paar Worte zu dieser Ausgabe des Newsletters"
 msgid "Latest news"
 msgstr "Aktuelles"
 
+msgid "Show news as tiles"
+msgstr "Aktuelles als Kacheln anzeigen"
+
+msgid ""
+"If checked, news are displayed as tiles. Otherwise, news are listed in full "
+"length."
+msgstr ""
+"Wenn diese Option aktiviert ist, werden die Nachrichten als Kacheln "
+"angezeigt. Anderfalls werden die Nachrichten in voller Länge "
+"untereinander aufgelistet."
+
 msgid "Events"
 msgstr "Veranstaltungen"
 
@@ -1061,8 +1072,8 @@ msgid ""
 "There are currently no allowed domains for iFrames. To enable domains for "
 "iFrames, please contact info@seantis.ch."
 msgstr ""
-"Derzeit sind keine Domains für iFrames zugelassen. Um Domains für iFrames "
-"zu aktivieren, kontaktieren Sie bitte info@seantis.ch"
+"Derzeit sind keine Domains für iFrames zugelassen. Um Domains für iFrames zu "
+"aktivieren, kontaktieren Sie bitte info@seantis.ch"
 
 msgid "Height"
 msgstr "Höhe"
@@ -2091,6 +2102,9 @@ msgstr "den ganzen Tag"
 msgid "Homepage"
 msgstr "Startseite"
 
+msgid "Save"
+msgstr "Speichern"
+
 msgid "Forms"
 msgstr "Formulare"
 
@@ -2523,6 +2537,20 @@ msgstr ""
 
 msgid "Sidebar links"
 msgstr "Links in der Sidebar"
+
+msgid "Delete content when expired"
+msgstr "Inhalt löschen wenn er abgelaufen ist"
+
+msgid ""
+"This content is automatically deleted when the end date has passed. If no "
+"end date is set, it will be deleted one day after start date."
+msgstr ""
+"Dieser Inhalt wird automatisch gelöscht, wenn das Enddatum überschritten "
+"ist. Wenn kein Enddatum festgelegt ist, wird er einen Tag nach dem "
+"Startdatum gelöscht."
+
+msgid "Delete content"
+msgstr "Inhalt löschen"
 
 msgid "This month"
 msgstr "Diesen Monat"
@@ -5690,15 +5718,3 @@ msgstr "Der Benutzer wurde erfolgreich erstellt"
 
 #~ msgid "Chat with"
 #~ msgstr "Chat mit"
-
-msgid "Delete content"
-msgstr "Inhalt löschen"
-
-msgid "Delete content when expired"
-msgstr "Inhalt löschen wenn er abgelaufen ist"
-
-msgid "This content is automatically deleted when the end date has passed. If "
-"no end date is set, it will be deleted one day after start date."
-msgstr "Dieser Inhalt wird automatisch gelöscht, wenn das Enddatum "
-"überschritten ist. Wenn kein Enddatum festgelegt ist, wird er einen Tag "
-"nach dem Startdatum gelöscht."
