+++ src/onegov/org/theme/styles/org.scss
@@ -1597,6 +1597,7 @@ $copy-link-icon: '\f0c5';
 $delete-link-icon: '\f014';
 $disabled-icon: '\f05e';
 $edit-link-icon: '\f040';
+$save-link-icon: '\f0c7';
 $export-link-icon: '\f019';
 $file-link-icon: '\f0c6';
 $image-link-icon: '\f03e';
@@ -1721,7 +1722,8 @@ $editbar-bg-color: $topbar-link-bg-hover;
         font-size: .8rem;
     }
 
-    a::before {
+    a::before,
+    button::before {
         display: inline-block;
         margin-right: .8ex;
         text-align: center;
@@ -1733,6 +1735,7 @@ $editbar-bg-color: $topbar-link-bg-hover;
     .copy-link::before { @include icon($copy-link-icon); }
     .delete-link::before { @include icon($delete-link-icon); }
     .edit-link::before { @include icon($edit-link-icon); }
+    .save-link::before { @include icon($save-link-icon); }
     .export-link::before { @include icon($export-link-icon); }
     .envelope::before { @include icon($envelope-icon); }
     .move-link::before { @include icon($move-link-icon); }
